requirements.txt
flask
requests
youtube_transcript_api
dotenv
pytest
google-genai
google-generativeai